Moving const and type defs to types for reuse
